+++ gammapy/datasets/map.py
