+++ gammapy/modeling/serialize.py
@@ -1,6 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Utilities to serialize models."""
-from gammapy.datasets.fit import MapDataset
+from gammapy.datasets import MapDataset
 from gammapy.spectrum import FluxPointsDataset, SpectrumDatasetOnOff
 from .models import Registry, SkyDiffuseCube, SkyModel
